Bugfix: Type mismatch in GD and Imagick drivers

+++ src/Jbarc/Barcode/ImageDriver/Driver.php
@@ -8,10 +8,10 @@ use Jbarc\Color\Color;
 
 interface Driver
 {
-    public function initPicture(float $width, float $height, Color $color);
+    public function initPicture(int $width, int $height, Color $color);
 
 
-    public function addRectangle(float $x1, float $y1, float $width, float $height);
+    public function addRectangle(int $x1, int $y1, int $width, int $height);
 
 
     public function getPicture();
